Finish RBAC in existing implementations
