--- backend/app/models/__init__.py
@@ -1,3 +1,0 @@
-from .user import User
-from .user import Chat
-from .user import Message
