+++ backend/app/routes/users.py
@@ -26,7 +26,7 @@ def load_moderator_page():
 
     return jsonify({'message': 'Authorized'})
 
-  
+
 @user_bp.route('profile', methods=['GET'])
 def get_current_user():
     authorization = is_user_authorized('user')
@@ -248,25 +248,6 @@ THIS HEADER DENOTES THAT THE FOLLOWING API ROUTE MEETS ONE OF THE FOLLOWING CRIT
 '''
 
 
-@user_bp.route('<int:user_id>/chats', methods=['POST'])
-def create_user_chat(user_id):
-    user = User.query.get(user_id)
-
-    if not user:
-        return jsonify({'error': 'User not found'}), 404
-
-
-    return jsonify({'message': 'Chat created successfully', 'chat_id': chat.id, 'user_id': user.id}), 201
-
-
-'''
-IMPORTANT:
-THIS HEADER DENOTES THAT THE FOLLOWING API ROUTE MEETS ONE OF THE FOLLOWING CRITERIA:
-- API ROUTE IS NEVER USED IN THE CLIENT-SIDE APPLICATION
-- API ROUTE NEEDS RBAC IMPLEMENTED IN IT IF NECESSARY (A USER DB CONNECTION PERFORMING ACTIONS ON THEIR BEHALF, NOT THE ADMIN DB CONNECTION)
-'''
-
-
 @user_bp.route('<int:user_id>/chats/<int:chat_id>', methods=['DELETE'])
 def delete_user_chat(user_id, chat_id):
     user = User.query.get(user_id)
@@ -288,17 +269,16 @@ def delete_user_chat(user_id, chat_id):
 
     db.session.execute(user_chat_association.delete().where(
         (user_chat_association.c.user_id == user_id) &
-            (user_chat_association.c.chat_id == chat_id)
+        (user_chat_association.c.chat_id == chat_id)
     ))
     db.session.commit()
 
     return jsonify({'message': 'Chat successfully removed from the user'}), 200
 
 
-
 @user_bp.route('<int:user_id>/chats/<int:chat_id>/messages', methods=['POST'])
 def create_message(user_id, chat_id):
-    authorization = is_user_authorized('user') 
+    authorization = is_user_authorized('user')
     if isinstance(authorization, tuple):
         return authorization
 
@@ -323,7 +303,8 @@ def create_message(user_id, chat_id):
         if 'content' not in message_json or not message_json['content'].strip():
             return jsonify({'error': 'Missing message content'})
 
-        message = Message(content=message_json['content'], chat_id=chat_id, sender_id=user_id)
+        message = Message(
+            content=message_json['content'], chat_id=chat_id, sender_id=user_id)
         user_db_session.add(message)
         user_db_session.commit()
 
@@ -352,20 +333,23 @@ def favorite_a_listing_for_user():
         if not listing:
             return jsonify({'error': 'Listing not found'}), 404
 
-
-        chat = user_db_session.query(Chat).filter_by(user_send=user.username, user_receive=listing.user.username).first()
+        chat = user_db_session.query(Chat).filter_by(
+            user_send=user.username, user_receive=listing.user.username).first()
 
         if not chat:
-            chat = Chat(user_send=user.username, user_receive=listing.user.username)
+            chat = Chat(user_send=user.username,
+                        user_receive=listing.user.username)
             user_db_session.add(chat)
             user_db_session.commit()
 
             # Create associations for both users
             new_association = {'user_id': user_id, 'chat_id': chat.id}
-            user_db_session.execute(user_chat_association.insert().values(new_association))
+            user_db_session.execute(
+                user_chat_association.insert().values(new_association))
 
             new_association = {'user_id': listing.user_id, 'chat_id': chat.id}
-            user_db_session.execute(user_chat_association.insert().values(new_association))
+            user_db_session.execute(
+                user_chat_association.insert().values(new_association))
 
             user_db_session.commit()
 
@@ -461,6 +445,41 @@ def pass_a_listing_for_user():
     return jsonify({'message': 'Listing successfully passed for the user'}), 201
 
 
+@user_bp.route('/favorited_chats', methods=['GET'])
+def get_chats():
+    authorization = is_user_authorized('user')
+    if isinstance(authorization, tuple):
+        return authorization
+
+    user_id = session['user_id']
+
+    with safe_db_connection(session.get('username'), session.get('password')) as user_db_session:
+        user = user_db_session.query(User).get(user_id)
+
+        favorited_listings = user.favorited_listings
+        owned_listings = user.listings
+
+        chats_data = []
+        visited_chats = set()
+
+        process_listings(favorited_listings, user, chats_data, visited_chats)
+        process_listings(owned_listings, user, chats_data, visited_chats)
+
+    return jsonify(chats_data)
+
+
+@user_bp.route('/messages/<int:chat_id>')
+def get_messages(chat_id):
+    authorization = is_user_authorized('user')
+    if isinstance(authorization, tuple):
+        return authorization
+
+    with safe_db_connection(session.get('username'), session.get('password')) as user_db_session:
+        messages = user_db_session.query(Message).filter_by(
+            chat_id=chat_id).order_by(Message.timestamp.asc()).all()
+    return jsonify([message.to_dict() for message in messages])
+
+
 def process_listings(listings, user, chats_data, visited_chats):
     for listing in listings:
         relevant_users = set()
@@ -477,43 +496,16 @@ def process_listings(listings, user, chats_data, visited_chats):
         for other_username in relevant_users:
             chats = Chat.query.filter(
                 ((Chat.user_send == user.username) & (Chat.user_receive == other_username)) |
-                ((Chat.user_receive == user.username) & (Chat.user_send == other_username))
+                ((Chat.user_receive == user.username)
+                 & (Chat.user_send == other_username))
             ).all()
             for chat in chats:
                 chat_key = tuple(sorted((user.username, other_username)))
                 if chat_key not in visited_chats:
-                    visited_chats.add(chat_key)  
+                    visited_chats.add(chat_key)
                     chats_data.append({
                         'chat_id': chat.id,
                         'other_user': other_username,
                         'listing_id': listing.id,
                         'listing_name': listing.name
                     })
-
-
-@user_bp.route('/favorited_chats', methods=['GET'])
-def get_chats():
-    if 'user_id' not in session:
-        return jsonify({'error': 'User not authenticated'}), 401
-
-    user_id = session['user_id']
-    user = User.query.get(user_id)
-    if not user:
-        return jsonify({'error': 'User not found'}), 404
-    
-    favorited_listings = user.favorited_listings
-    owned_listings = user.listings 
-
-    chats_data = []
-    visited_chats = set()
-
-    process_listings(favorited_listings, user, chats_data, visited_chats)
-    process_listings(owned_listings, user, chats_data, visited_chats)
-
-    return jsonify(chats_data)
-
-
-@user_bp.route('/messages/<int:chat_id>')
-def get_messages(chat_id):
-    messages = Message.query.filter_by(chat_id=chat_id).order_by(Message.timestamp.asc()).all()
-    return jsonify([message.to_dict() for message in messages])
